# Fireblocks.swagger.yml
...
openapi: 3.0.0
info:
  title: Fireblocks API
  version: "1.5.0"
servers:
  - url: 'https://api.fireblocks.io/v1'
paths:
...
  '/vault/accounts/{vaultAccountId}/{assetId}/addresses':
    get:
      summary: Returns all addresses generated for a Vault wallet
      parameters:
        - in: path
          name: vaultAccountId
          required: true
          description: >-
            The ID of the vault account to return
          schema:
            type: string
        - in: path
          name: assetId
          required: true
          description: The ID of the asset
          schema:
            type: string
      responses:
        '200':
          description: A list of deposit addresses
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VaultWalletAddress'
        default:
          $ref: '#/components/responses/Error'
    post:
      summary: Generate a new deposit address for a Vault wallet
      requestBody:
        required: false
        content:
          application/json:
            schema:
              properties:
                description:
                  description: (Optional) Attach a description to the new address
                  type: string
                customerRefId:
                  description: >-
                    Optional - Sets a customer reference ID
                  type: string
      parameters:
        - in: path
          name: vaultAccountId
          required: true
          description: >-
            The ID of the vault account to return
          schema:
            type: string
        - in: path
          name: assetId
          required: true
          description: The ID of the asset
          schema:
            type: string
      responses:
        '200':
          description: The created address
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CreateAddressResponse'
        default:
          $ref: '#/components/responses/Error'
...
security:
  - bearerTokenAuth: [ ]
  - ApiKeyAuth: [ ]
components:
  responses:
    Error:
      description: Error Response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerTokenAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    VaultWalletAddress:
      type: object
      properties:
        assetId:
          type: string
        address:
          type: string
        description:
          type: string
        tag:
          type: string
        type:
          type: string
        customerRefId:
          type: string
        addressFormat:
          type: string
          enum:
            - SEGWIT
            - LEGACY
            - BASE
            - PAYMENT
        legacyAddress:
          type: string
        enterpriseAddress:
          type: string
        bip44AddressIndex:
          type: integer
    CreateAddressResponse:
      type: object
      properties:
        address:
          type: string
        legacyAddress:
          type: string
        enterpriseAddress:
          type: string
        tag:
          type: string
        bip44AddressIndex:
          type: integer
...
    Error:
      type: object
      properties:
        message:
          type: string
        code:
          type: number
          format: decimal
...